+++ custom_components/hikvision_next/api/const.py
@@ -1,0 +1,14 @@
+GET = "GET"
+PUT = "PUT"
+POST = "POST"
+
+CONNECTION_TYPE_DIRECT = "Direct"
+CONNECTION_TYPE_PROXIED = "Proxied"
+
+EVENTS_ALTERNATE_ID = {
+    "vmd": "motiondetection",
+    "shelteralarm": "tamperdetection",
+    "VMDHumanVehicle": "motiondetection",
+}
+
+MUTEX_ALTERNATE_ID = {"motiondetection": "VMDHumanVehicle"}
