+++ custom_components/hikvision_next/config_flow.py
@@ -16,7 +16,7 @@ from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_USERNAME
 from homeassistant.data_entry_flow import FlowResult
 from homeassistant.helpers.httpx_client import get_async_client
 
-from .const import DATA_ALARM_SERVER_HOST, DATA_SET_ALARM_SERVER, DOMAIN
+from .const import CONF_ALARM_SERVER_HOST, CONF_SET_ALARM_SERVER, DOMAIN
 from .isapi import ISAPI
 
 _LOGGER = logging.getLogger(__name__)
@@ -38,12 +38,12 @@ class HikvisionFlowHandler(ConfigFlow, domain=DOMAIN):
                 vol.Required(CONF_USERNAME, default=user_input.get(CONF_USERNAME, "")): str,
                 vol.Required(CONF_PASSWORD, default=user_input.get(CONF_PASSWORD, "")): str,
                 vol.Required(
-                    DATA_SET_ALARM_SERVER,
-                    default=user_input.get(DATA_SET_ALARM_SERVER, True),
+                    CONF_SET_ALARM_SERVER,
+                    default=user_input.get(CONF_SET_ALARM_SERVER, True),
                 ): bool,
                 vol.Required(
-                    DATA_ALARM_SERVER_HOST,
-                    default=user_input.get(DATA_ALARM_SERVER_HOST, f"http://{local_ip}:8123"),
+                    CONF_ALARM_SERVER_HOST,
+                    default=user_input.get(CONF_ALARM_SERVER_HOST, f"http://{local_ip}:8123"),
                 ): str,
             }
         )
