+++ custom_components/hikvision_next/const.py
@@ -6,17 +6,13 @@ from homeassistant.components.binary_sensor import BinarySensorDeviceClass
 
 DOMAIN: Final = "hikvision_next"
 
-DATA_SET_ALARM_SERVER: Final = "set_alarm_server"
-DATA_ALARM_SERVER_HOST: Final = "alarm_server"
-DATA_ISAPI = "isapi"
+CONF_SET_ALARM_SERVER: Final = "set_alarm_server"
+CONF_ALARM_SERVER_HOST: Final = "alarm_server"
 ALARM_SERVER_PATH = "/api/hikvision"
 
 EVENTS_COORDINATOR: Final = "events"
 SECONDARY_COORDINATOR: Final = "secondary"
 HOLIDAY_MODE = "holiday_mode"
-
-CONNECTION_TYPE_DIRECT = "Direct"
-CONNECTION_TYPE_PROXIED = "Proxied"
 
 ATTR_CONFIG_ENTRY_ID = "config_entry_id"
 ACTION_REBOOT = "reboot"
@@ -97,14 +93,6 @@ EVENTS = {
         "device_class": BinarySensorDeviceClass.MOTION,
     }
 }
-
-EVENTS_ALTERNATE_ID = {
-    "vmd": "motiondetection",
-    "shelteralarm": "tamperdetection",
-    "VMDHumanVehicle": "motiondetection",
-}
-
-MUTEX_ALTERNATE_IDS = {"motiondetection": "VMDHumanVehicle"}
 
 STREAM_TYPE = {
     1: "Main Stream",
